Makefile flag update
- Update Makefile to use conversion warning flag
- Fixed resulting warnings
- refprop and json_library have new flag to ignore warning as they are external libraries

+++ machine_build.f90
@@ -293,7 +293,8 @@ contains
           fdiags(1) = wwp1*tfno/rmajor
           call report_error(141)
        else if (ripflag == 2) then
-          idiags(1) = tfno
+          ! Convert to integer as idiags is integer array
+          idiags(1) = INT(tfno)
           call report_error(142)
        else
           fdiags(1) = (rmajor+rminor)/rtot
